pyproject.toml
[project]
name = "microservice-payment"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiokafka>=0.13.0",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "pre-commit>=4.5.1",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "ruff>=0.15.0",
    "tenacity>=9.1.4",
]
